Changed game id to long type. Changed game submission to GET instead of POST. Altered id attributes of score display to be in English. Added name attributes to scoring display elements.

+++ src/main/java/com/perscholas/scorekeeper/dao/GameDAO.java
@@ -10,6 +10,6 @@ import java.util.List;
 @Repository
 public interface GameDAO extends JpaRepository<Game, Long> {
 	List<Game> findAll();
-	Game findById(int id);
+	Game findById(long id);
 	Game findByPlayers(Player player);
 }
